+++ stripe/api_resources/customer.py
@@ -6,6 +6,7 @@ from stripe import util
 from stripe.api_resources.abstract import CreateableAPIResource
 from stripe.api_resources.abstract import DeletableAPIResource
 from stripe.api_resources.abstract import ListableAPIResource
+from stripe.api_resources.abstract import SearchableAPIResource
 from stripe.api_resources.abstract import UpdateableAPIResource
 from stripe.api_resources.abstract import custom_method
 from stripe.api_resources.abstract import nested_resource_class_methods
@@ -13,9 +14,19 @@ from stripe.api_resources.abstract import nested_resource_class_methods
 
 @custom_method("delete_discount", http_verb="delete", http_path="discount")
 @custom_method(
+    "create_funding_instructions",
+    http_verb="post",
+    http_path="funding_instructions",
+)
+@custom_method(
     "list_payment_methods",
     http_verb="get",
     http_path="payment_methods",
+)
+@nested_resource_class_methods(
+    "cash_balance",
+    operations=["retrieve", "update"],
+    resource_plural="cash_balance",
 )
 @nested_resource_class_methods(
     "balance_transaction",
@@ -33,9 +44,17 @@ class Customer(
     CreateableAPIResource,
     DeletableAPIResource,
     ListableAPIResource,
+    SearchableAPIResource,
     UpdateableAPIResource,
 ):
     OBJECT_NAME = "customer"
+
+    def create_funding_instructions(self, idempotency_key=None, **params):
+        url = self.instance_url() + "/funding_instructions"
+        headers = util.populate_headers(idempotency_key)
+        resp = self.request("post", url, params, headers)
+        stripe_object = util.convert_to_stripe_object(resp)
+        return stripe_object
 
     def list_payment_methods(self, idempotency_key=None, **params):
         url = self.instance_url() + "/payment_methods"
@@ -44,6 +63,14 @@ class Customer(
         stripe_object = util.convert_to_stripe_object(resp)
         stripe_object._retrieve_params = params
         return stripe_object
+
+    @classmethod
+    def search(cls, *args, **kwargs):
+        return cls._search(search_url="/v1/customers/search", *args, **kwargs)
+
+    @classmethod
+    def search_auto_paging_iter(cls, *args, **kwargs):
+        return cls.search(*args, **kwargs).auto_paging_iter()
 
     def delete_discount(self, **params):
         requestor = api_requestor.APIRequestor(