+++ tests/test_generated_examples.py
@@ -122,8 +122,7 @@ class TestGeneratedExamples(object):
 
     def test_filelink_update(self, request_mock):
         stripe.FileLink.modify(
-            "link_xxxxxxxxxxxxx",
-            metadata={"order_id": "6735"},
+            "link_xxxxxxxxxxxxx", metadata={"order_id": "6735"}
         )
         request_mock.assert_requested(
             "post", "/v1/file_links/link_xxxxxxxxxxxxx"
@@ -504,8 +503,7 @@ class TestGeneratedExamples(object):
 
     def test_creditnote_update(self, request_mock):
         stripe.CreditNote.modify(
-            "cn_xxxxxxxxxxxxx",
-            metadata={"order_id": "6735"},
+            "cn_xxxxxxxxxxxxx", metadata={"order_id": "6735"}
         )
         request_mock.assert_requested(
             "post", "/v1/credit_notes/cn_xxxxxxxxxxxxx"
@@ -663,8 +661,7 @@ class TestGeneratedExamples(object):
 
     def test_invoiceitem_update(self, request_mock):
         stripe.InvoiceItem.modify(
-            "ii_xxxxxxxxxxxxx",
-            metadata={"order_id": "6735"},
+            "ii_xxxxxxxxxxxxx", metadata={"order_id": "6735"}
         )
         request_mock.assert_requested(
             "post", "/v1/invoiceitems/ii_xxxxxxxxxxxxx"
@@ -1405,53 +1402,6 @@ class TestGeneratedExamples(object):
         stripe.terminal.Reader.list(limit=3)
         request_mock.assert_requested("get", "/v1/terminal/readers")
 
-    def test_order_create(self, request_mock):
-        stripe.Order.create(
-            currency="usd",
-            email="jenny.rosen@example.com",
-            items=[{"type": "sku", "parent": "sku_xxxxxxxxxxxxx"}],
-            shipping={
-                "name": "Jenny Rosen",
-                "address": {
-                    "line1": "1234 Main Street",
-                    "city": "San Francisco",
-                    "state": "CA",
-                    "country": "US",
-                    "postal_code": "94111",
-                },
-            },
-        )
-        request_mock.assert_requested("post", "/v1/orders")
-
-    def test_order_retrieve(self, request_mock):
-        stripe.Order.retrieve("or_xxxxxxxxxxxxx")
-        request_mock.assert_requested("get", "/v1/orders/or_xxxxxxxxxxxxx")
-
-    def test_order_update(self, request_mock):
-        stripe.Order.modify("or_xxxxxxxxxxxxx", metadata={"order_id": "6735"})
-        request_mock.assert_requested("post", "/v1/orders/or_xxxxxxxxxxxxx")
-
-    def test_order_pay(self, request_mock):
-        stripe.Order.pay("or_xxxxxxxxxxxxx", source="tok_xxxx")
-        request_mock.assert_requested(
-            "post", "/v1/orders/or_xxxxxxxxxxxxx/pay"
-        )
-
-    def test_order_list(self, request_mock):
-        stripe.Order.list(limit=3)
-        request_mock.assert_requested("get", "/v1/orders")
-
-    def test_orderreturn_retrieve(self, request_mock):
-        stripe.OrderReturn.retrieve("orret_xxxxxxxxxxxxx")
-        request_mock.assert_requested(
-            "get",
-            "/v1/order_returns/orret_xxxxxxxxxxxxx",
-        )
-
-    def test_orderreturn_list(self, request_mock):
-        stripe.OrderReturn.list(limit=3)
-        request_mock.assert_requested("get", "/v1/order_returns")
-
     def test_sku_create(self, request_mock):
         stripe.SKU.create(
             attributes={"size": "Medium", "gender": "Unisex"},
@@ -1625,3 +1575,170 @@ class TestGeneratedExamples(object):
     def test_paymentlink_retrieve(self, request_mock):
         stripe.PaymentLink.retrieve("pl_xyz")
         request_mock.assert_requested("get", "/v1/payment_links/pl_xyz")
+
+    def test_paymentintent_verify_microdeposits(self, request_mock):
+        stripe.PaymentIntent.verify_microdeposits("pi_xxxxxxxxxxxxx")
+        request_mock.assert_requested(
+            "post",
+            "/v1/payment_intents/pi_xxxxxxxxxxxxx/verify_microdeposits",
+        )
+
+    def test_setupintent_verify_microdeposits(self, request_mock):
+        stripe.SetupIntent.verify_microdeposits("seti_xxxxxxxxxxxxx")
+        request_mock.assert_requested(
+            "post",
+            "/v1/setup_intents/seti_xxxxxxxxxxxxx/verify_microdeposits",
+        )
+
+    def test_test_helpers_testclock_create(self, request_mock):
+        stripe.test_helpers.TestClock.create(frozen_time=123, name="cogsworth")
+        request_mock.assert_requested("post", "/v1/test_helpers/test_clocks")
+
+    def test_test_helpers_testclock_retrieve(self, request_mock):
+        stripe.test_helpers.TestClock.retrieve("clock_xyz")
+        request_mock.assert_requested(
+            "get",
+            "/v1/test_helpers/test_clocks/clock_xyz",
+        )
+
+    def test_test_helpers_testclock_list(self, request_mock):
+        stripe.test_helpers.TestClock.list()
+        request_mock.assert_requested("get", "/v1/test_helpers/test_clocks")
+
+    def test_test_helpers_testclock_delete(self, request_mock):
+        stripe.test_helpers.TestClock.delete("clock_xyz")
+        request_mock.assert_requested(
+            "delete",
+            "/v1/test_helpers/test_clocks/clock_xyz",
+        )
+
+    def test_test_helpers_testclock_advance(self, request_mock):
+        stripe.test_helpers.TestClock.advance("clock_xyz", frozen_time=142)
+        request_mock.assert_requested(
+            "post",
+            "/v1/test_helpers/test_clocks/clock_xyz/advance",
+        )
+
+    def test_customer_create_funding_instructions(self, request_mock):
+        stripe.Customer.create_funding_instructions(
+            "cus_123",
+            bank_transfer={
+                "requested_address_types": ["zengin"],
+                "type": "jp_bank_transfer",
+            },
+            currency="usd",
+            funding_type="bank_transfer",
+        )
+        request_mock.assert_requested(
+            "post",
+            "/v1/customers/cus_123/funding_instructions",
+        )
+
+    def test_terminal_configuration_list(self, request_mock):
+        stripe.terminal.Configuration.list()
+        request_mock.assert_requested("get", "/v1/terminal/configurations")
+
+    def test_terminal_configuration_retrieve(self, request_mock):
+        stripe.terminal.Configuration.retrieve("uc_123")
+        request_mock.assert_requested(
+            "get", "/v1/terminal/configurations/uc_123"
+        )
+
+    def test_terminal_configuration_create(self, request_mock):
+        stripe.terminal.Configuration.create()
+        request_mock.assert_requested("post", "/v1/terminal/configurations")
+
+    def test_terminal_configuration_update(self, request_mock):
+        stripe.terminal.Configuration.modify(
+            "uc_123",
+            tipping={"usd": {"fixed_amounts": [10]}},
+        )
+        request_mock.assert_requested(
+            "post", "/v1/terminal/configurations/uc_123"
+        )
+
+    def test_terminal_configuration_delete(self, request_mock):
+        stripe.terminal.Configuration.delete("uc_123")
+        request_mock.assert_requested(
+            "delete",
+            "/v1/terminal/configurations/uc_123",
+        )
+
+    def test_refund_expire(self, request_mock):
+        stripe.Refund.TestHelpers.expire("re_123")
+        request_mock.assert_requested(
+            "post",
+            "/v1/test_helpers/refunds/re_123/expire",
+        )
+
+    def test_order_create(self, request_mock):
+        stripe.Order.create(
+            description="description",
+            currency="usd",
+            line_items=[{"description": "my line item"}],
+        )
+        request_mock.assert_requested("post", "/v1/orders")
+
+    def test_order_update(self, request_mock):
+        stripe.Order.modify("order_xyz")
+        request_mock.assert_requested("post", "/v1/orders/order_xyz")
+
+    def test_order_list_line_items(self, request_mock):
+        stripe.Order.list_line_items("order_xyz")
+        request_mock.assert_requested("get", "/v1/orders/order_xyz/line_items")
+
+    def test_order_cancel(self, request_mock):
+        stripe.Order.cancel("order_xyz")
+        request_mock.assert_requested("post", "/v1/orders/order_xyz/cancel")
+
+    def test_order_reopen(self, request_mock):
+        stripe.Order.reopen("order_xyz")
+        request_mock.assert_requested("post", "/v1/orders/order_xyz/reopen")
+
+    def test_order_submit(self, request_mock):
+        stripe.Order.submit("order_xyz", expected_total=100)
+        request_mock.assert_requested("post", "/v1/orders/order_xyz/submit")
+
+    def test_order_update2(self, request_mock):
+        stripe.Order.modify("order_xyz")
+        request_mock.assert_requested("post", "/v1/orders/order_xyz")
+
+    def test_financial_connections_account_retrieve(self, request_mock):
+        stripe.financial_connections.Account.retrieve("fca_xyz")
+        request_mock.assert_requested(
+            "get",
+            "/v1/financial_connections/accounts/fca_xyz",
+        )
+
+    def test_financial_connections_account_refresh_account(self, request_mock):
+        stripe.financial_connections.Account.refresh_account(
+            "fca_xyz",
+            features=["balance"],
+        )
+        request_mock.assert_requested(
+            "post",
+            "/v1/financial_connections/accounts/fca_xyz/refresh",
+        )
+
+    def test_financial_connections_account_disconnect(self, request_mock):
+        stripe.financial_connections.Account.disconnect("fca_xyz")
+        request_mock.assert_requested(
+            "post",
+            "/v1/financial_connections/accounts/fca_xyz/disconnect",
+        )
+
+    def test_financial_connections_session_create(self, request_mock):
+        stripe.financial_connections.Session.create(
+            account_holder={"type": "customer", "customer": "cus_123"},
+            permissions=["balances"],
+        )
+        request_mock.assert_requested(
+            "post", "/v1/financial_connections/sessions"
+        )
+
+    def test_financial_connections_session_retrieve(self, request_mock):
+        stripe.financial_connections.Session.retrieve("fcsess_xyz")
+        request_mock.assert_requested(
+            "get",
+            "/v1/financial_connections/sessions/fcsess_xyz",
+        )